Pre-refactoring completed Rhino code. This will move about once I've committed the CSP and CFG code, but I'm getting nervous about keeping all of this only locally.
Also in dire need of more Javadocs, particularly in the API. Test coverage isn't bad, though. Includes both the small .api classes and the (much bigger) Rhino implementation thereof, to allow us to migrate from Rhino (I increasingly think we want to do that at some point in the distant future).

The one major outstanding thing codewise is that we don't exploit securitycontrollers, etc, which are available to us, and for belt-and-braces, we really should.